Cleaned up the main file
moved everthing into the library
