+++ PythonForCounting/Libraries/Thresholding.py
@@ -11,3 +11,35 @@ def getColorHistogramFromSlice(slice):
         outputBinEdges.append(binEdges)
     return outputHistograms, outputBinEdges
 
+def calculateIntensity(pixel):
+    """
+    Returnerer intensiteten af en enkelt pixel
+    :param pixel:
+    :return:
+    """
+    bgrMean = pixel[0] / 3 + pixel[1] / 3 + pixel[2] / 3
+    intensity = bgrMean / 255
+    return intensity
+
+
+def makeImageBinaryIntensityThreshold(img, threshold):
+    """
+    :param img:
+    :param threshold:
+    :return black and white output:
+
+    Funktion der gør pixels med en intensity værdi større end threshold sorte
+    og pixels med en intensity værdi mindre end threshold hvide
+    Skal tage imod et bgr image
+
+    Linus er den bedste 😎
+    -------
+    """
+    output = np.zeros((img.shape[0], img.shape[1]), dtype=np.uint8)
+    for y, row in enumerate(img):
+        for x, pixel in enumerate(row):
+            if calculateIntensity(img[y, x]) < threshold:
+                output[y, x] = 255
+            else:
+                output[y, x] = 0
+    return output
